Refactor, some pvp, stagger with profitables
Had to move away from boolean [item] to item [int] and such to enable
conditional modification to some of the data entries for things like
some things you might want to do in pvp, and in the future will probably
be used to support banishes with complicated conditions (like Unleash
Nanites). Additionally, when manually using a non-staggering item (like
abstractions in  the DMT), a staggering item will be paired with it if
possible.